# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TalentRank API
  version: 0.1.0
  description: Human-in-the-loop candidate ranking API. All tenant resources are workspace scoped.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      security: []
      tags: [Auth]
      operationId: register
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RegisterRequest" } } } }
      responses:
        "202": { description: Registration accepted; email verification is pending, content: { application/json: { schema: { $ref: "#/components/schemas/VerificationPending" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /auth/login:
    post:
      security: []
      tags: [Auth]
      operationId: login
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/LoginRequest" } } } }
      responses:
        "200": { description: Authenticated, content: { application/json: { schema: { $ref: "#/components/schemas/AuthTokens" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/refresh:
    post:
      security: [{ refreshCookie: [] }]
      tags: [Auth]
      operationId: refresh
      responses:
        "200": { description: Rotated tokens, content: { application/json: { schema: { $ref: "#/components/schemas/AuthTokens" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/verify-email:
    post:
      security: []
      tags: [Auth]
      operationId: verifyEmail
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/TokenRequest" } } } }
      responses:
        "200": { description: Email verified and session created, content: { application/json: { schema: { $ref: "#/components/schemas/AuthTokens" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/resend-verification:
    post:
      security: []
      tags: [Auth]
      operationId: resendVerification
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/EmailRequest" } } } }
      responses:
        "202": { description: Accepted whether or not an eligible account exists, content: { application/json: { schema: { $ref: "#/components/schemas/VerificationPending" } } } }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/invitations/inspect:
    post:
      security: []
      tags: [Auth]
      operationId: inspectInvitation
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/TokenRequest" } } } }
      responses:
        "200": { description: Invitation details, content: { application/json: { schema: { $ref: "#/components/schemas/InvitationInspection" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/invitations/accept:
    post:
      security: []
      tags: [Auth]
      operationId: acceptInvitation
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AcceptInvitationRequest" } } } }
      responses:
        "200": { description: Invitation accepted and session created, content: { application/json: { schema: { $ref: "#/components/schemas/AuthTokens" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /auth/switch-workspace:
    post:
      tags: [Auth]
      operationId: switchWorkspace
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [workspace_id], properties: { workspace_id: { type: string, format: uuid } } } } } }
      responses:
        "200": { description: New workspace-scoped session, content: { application/json: { schema: { $ref: "#/components/schemas/AuthTokens" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /workspace/members/{memberId}:
    parameters: [{ $ref: "#/components/parameters/MemberId" }]
    patch:
      tags: [Workspace]
      operationId: updateWorkspaceMember
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/MemberUpdate" } } } }
      responses:
        "200": { description: Member updated, content: { application/json: { schema: { $ref: "#/components/schemas/WorkspaceMember" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Workspace]
      operationId: removeWorkspaceMember
      responses:
        "204": { description: Membership removed }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /workspace/invitations:
    get:
      tags: [Workspace]
      operationId: listWorkspaceInvitations
      responses:
        "200": { description: Workspace invitations }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Workspace]
      operationId: createWorkspaceInvitation
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/InvitationWrite" } } } }
      responses:
        "201": { description: Invitation created, content: { application/json: { schema: { $ref: "#/components/schemas/Invitation" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /workspace/invitations/{invitationId}/resend:
    parameters: [{ $ref: "#/components/parameters/InvitationId" }]
    post:
      tags: [Workspace]
      operationId: resendWorkspaceInvitation
      responses:
        "202": { description: Invitation resent with a new token }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /workspace/invitations/{invitationId}:
    parameters: [{ $ref: "#/components/parameters/InvitationId" }]
    delete:
      tags: [Workspace]
      operationId: revokeWorkspaceInvitation
      responses:
        "204": { description: Invitation revoked }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    refreshCookie: { type: apiKey, in: cookie, name: talentrank_refresh }
  parameters:
...
    MemberId: { in: path, name: memberId, required: true, schema: { type: string, format: uuid } }
    InvitationId: { in: path, name: invitationId, required: true, schema: { type: string, format: uuid } }
  responses:
    Unauthorized: { description: Missing or invalid authentication, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    Forbidden: { description: Authenticated but not allowed, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    Conflict: { description: State conflict such as last_admin or token replay, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds before retry }
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    ValidationError: { description: Invalid request, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
...
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        request_id: { type: string }
        errors: { type: array, items: { type: object, additionalProperties: true } }
    RegisterRequest:
      type: object
      required: [email, password, display_name, workspace_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12, maxLength: 128 }
        display_name: { type: string, minLength: 2, maxLength: 120 }
        workspace_name: { type: string, minLength: 2, maxLength: 120 }
    VerificationPending:
      type: object
      required: [status]
      properties:
        status: { type: string, const: verification_pending }
        email: { type: string, format: email }
    EmailRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    TokenRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 32, writeOnly: true }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1, maxLength: 128 }
        remember_me: { type: boolean, default: false }
    AuthTokens:
      type: object
      required: [access_token, token_type, expires_in, user, workspace, role]
      properties:
        access_token: { type: string, writeOnly: true }
        token_type: { type: string, const: Bearer }
        expires_in: { type: integer, const: 900 }
        user: { $ref: "#/components/schemas/User" }
        workspace: { $ref: "#/components/schemas/Workspace" }
        role: { $ref: "#/components/schemas/Role" }
    User:
      type: object
      required: [id, email, display_name]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        email_verified_at: { type: [string, "null"], format: date-time }
    Role:
      type: string
      enum: [admin, recruiter, hiring_manager]
    Workspace:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
...
    WorkspaceMember:
      type: object
      required: [id, user, role, status, created_at]
      properties:
        id: { type: string, format: uuid }
        user: { $ref: "#/components/schemas/User" }
        role: { $ref: "#/components/schemas/Role" }
        status: { type: string, enum: [active, suspended] }
        last_accessed_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
    MemberUpdate:
      type: object
      minProperties: 1
      properties:
        role: { $ref: "#/components/schemas/Role" }
        status: { type: string, enum: [active, suspended] }
    InvitationWrite:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
    Invitation:
      allOf:
        - { $ref: "#/components/schemas/InvitationWrite" }
        - type: object
          required: [id, status, expires_at, created_at]
          properties:
            id: { type: string, format: uuid }
            status: { type: string, enum: [pending, expired, accepted, revoked] }
            expires_at: { type: string, format: date-time }
            created_at: { type: string, format: date-time }
            accepted_at: { type: [string, "null"], format: date-time }
            invited_by_name: { type: string }
    InvitationInspection:
      type: object
      required: [email, role, workspace, expires_at, account_exists]
      properties:
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        workspace: { $ref: "#/components/schemas/Workspace" }
        expires_at: { type: string, format: date-time }
        account_exists: { type: boolean }
    AcceptInvitationRequest:
      allOf:
        - { $ref: "#/components/schemas/TokenRequest" }
        - type: object
          properties:
            display_name: { type: string, minLength: 2, maxLength: 120 }
            password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
...